fix: decoding

+++ receiver/algos/hamming.odin
@@ -5,7 +5,7 @@ import "core:fmt"
 import "core:os"
 import "core:strconv"
 
-hamming_decode :: proc(input_length: uint, data_length: uint, encoded_input: u64) -> (output: u64, err_position: u64, redundant_bits_mask: u64) {
+hamming_decode :: proc(input_length: uint, data_length: uint, encoded_input: u64) -> (output: u64, err_position: uint, redundant_bits_mask: u64) {
 	r := input_length - data_length
 	for i: uint = 1; i <= r; i+=1 {
 		pos_mask:uint = 1<<(i-1)
@@ -30,12 +30,12 @@ hamming_decode :: proc(input_length: uint, data_length: uint, encoded_input: u64
 }
 
 @(private)
-extract_content :: proc (input_length: uint, redundant_bits_mask: uint, encoded_input: uint) -> uint {
-	output: uint = 0
+extract_content :: proc (input_length: uint, redundant_bits_mask: u64, encoded_input: u64) -> u64 {
+	output: u64 = 0
 	added_bit_pos_tracker: uint = 0
 
 	for i: uint = 0; i < input_length; i+=1 {
-		pos_mask:uint = 1 <<i
+		pos_mask:u64 = 1 <<i
 		if redundant_bits_mask & pos_mask > 0 {
 			continue
 		}
@@ -50,7 +50,7 @@ extract_content :: proc (input_length: uint, redundant_bits_mask: uint, encoded_
 }
 
 @(private)
-get_bytes_with :: proc(bit_to_flip_count: uint, encoded_input: uint, length: uint) -> (uint, uint) {
+get_bytes_with :: proc(bit_to_flip_count: uint, encoded_input: u64, length: uint) -> (uint, uint) {
 	flip_count := 0
 	bit_count: uint = 0
 	is_one := false
@@ -62,7 +62,7 @@ get_bytes_with :: proc(bit_to_flip_count: uint, encoded_input: uint, length: uin
 		// fmt.fprintf(os.stderr, "Analizing bit: %2d, IsOne: %t\n", i+1,is_one)
 
 		if is_one {
-			pos_mask: uint = 1<<(i-1)
+			pos_mask: u64 = 1<<(i-1)
 			// fmt.fprintf(os.stderr, "Bit is 1! PosMask: %b, Input: %b, PosMask & Input: %b\n", pos_mask, encoded_input, pos_mask & encoded_input)
 			if pos_mask & encoded_input > 0 {
 				ones_count += 1
